requirements.txt
streamlit
tensorflow==2.15.0
numpy
pillow
matplotlib
pandas
scikit-learn
reportlab
opencv-python-headless